setup api, web routes
