+++ app/routes/api.go
@@ -9,9 +9,5 @@ import (
 func apiRoutes() *bone.Mux {
 	router := bone.New()
 
-	router.GetFunc("/login", api.AuthMainHandler)
-	router.GetFunc("/o/auth/login/github", api.AuthGithubHandler)
-	router.GetFunc("/o/auth/callback/github", api.AuthGithubCallbackHandler)
-
 	return router
 }
