+++ ui/index.html
@@ -1,6 +1,6 @@
 <!DOCTYPE html>
 <head>
-    <base href="/">
+    <base href="/app">
     <title>Brizo</title>
     <link href="{{ assets.App["css"] }}" rel="stylesheet" />
 </head>
